Extract analyse module. Can refine later
