+++ compiler-core/src/javascript/tests.rs
@@ -38,7 +38,7 @@ macro_rules! assert_js {
         let _ = modules.insert("gleam".into(), $crate::type_::build_prelude(&ids));
         let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name.into();
-        let dep = $crate::type_::infer_module(
+        let dep = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -51,7 +51,7 @@ macro_rules! assert_js {
         let _ = modules.insert($dep_name.into(), dep.type_info);
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -77,7 +77,7 @@ macro_rules! assert_js {
         let _ = modules.insert("gleam".into(), $crate::type_::build_prelude(&ids));
         let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name.into();
-        let dep = $crate::type_::infer_module(
+        let dep = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -90,7 +90,7 @@ macro_rules! assert_js {
         let _ = modules.insert($dep_name.into("/"), dep.type_info);
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -119,7 +119,7 @@ macro_rules! assert_js {
 
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -147,7 +147,7 @@ macro_rules! assert_js {
 
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -177,7 +177,7 @@ macro_rules! assert_ts_def {
         let _ = modules.insert("gleam".into(), $crate::type_::build_prelude(&ids));
         let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name.into();
-        let dep = $crate::type_::infer_module(
+        let dep = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -190,7 +190,7 @@ macro_rules! assert_ts_def {
         let _ = modules.insert($dep_name.into(), dep.type_info);
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -215,7 +215,7 @@ macro_rules! assert_ts_def {
         let _ = modules.insert("gleam".into(), $crate::type_::build_prelude(&ids));
         let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name.into();
-        let dep = $crate::type_::infer_module(
+        let dep = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -228,7 +228,7 @@ macro_rules! assert_ts_def {
         let _ = modules.insert($dep_name.into("/"), dep.type_info);
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -256,7 +256,7 @@ macro_rules! assert_ts_def {
 
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
@@ -283,7 +283,7 @@ macro_rules! assert_ts_def {
 
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = "my/mod".into();
-        let ast = $crate::type_::infer_module(
+        let ast = $crate::analyse::infer_module(
             $crate::build::Target::JavaScript,
             &ids,
             ast,
