+++ compiler-core/src/type_/expression.rs
@@ -2,13 +2,16 @@ use itertools::Itertools;
 use vec1::Vec1;
 
 use super::{pipe::PipeTyper, *};
-use crate::ast::{
-    Arg, AssignName, AssignmentKind, BinOp, BitStringSegment, BitStringSegmentOption, CallArg,
-    Clause, ClauseGuard, Constant, HasLocation, RecordUpdateSpread, SrcSpan, TodoKind, TypeAst,
-    TypedArg, TypedClause, TypedClauseGuard, TypedConstant, TypedExpr, TypedMultiPattern,
-    UntypedArg, UntypedClause, UntypedClauseGuard, UntypedConstant,
-    UntypedConstantBitStringSegment, UntypedExpr, UntypedExprBitStringSegment, UntypedMultiPattern,
-    UntypedPattern, Use,
+use crate::{
+    analyse::infer_bit_string_segment_option,
+    ast::{
+        Arg, AssignName, AssignmentKind, BinOp, BitStringSegment, BitStringSegmentOption, CallArg,
+        Clause, ClauseGuard, Constant, HasLocation, RecordUpdateSpread, SrcSpan, TodoKind, TypeAst,
+        TypedArg, TypedClause, TypedClauseGuard, TypedConstant, TypedExpr, TypedMultiPattern,
+        UntypedArg, UntypedClause, UntypedClauseGuard, UntypedConstant,
+        UntypedConstantBitStringSegment, UntypedExpr, UntypedExprBitStringSegment,
+        UntypedMultiPattern, UntypedPattern, Use,
+    },
 };
 
 use im::hashmap;
@@ -1581,7 +1584,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         })
     }
 
-    // TODO: extract the type annotation checking into a infer_module_const
+    // TODO: extract the type annotation checking into a crate::analyse::infer_module_const
     // function that uses this function internally
     pub fn infer_const(
         &mut self,
